fix: apply feedback
